fix(model): bump `SessionStartLimit` field sizes (#2286)
The [largest
bot](https://gist.github.com/advaith1/451dcbca2d7c3503d4f48d63eb918cb0)
is in ~21M guilds, meaning that its session start limit := max(2000,
(guild_count / 1000) * 3) is just under u16::MAX (65535), so to
safeguard against overflows we should bump the field sizes a step (u16
-> u32).

I also went ahead and bumped the `max_concurrency` field to u16, size
it's plausible that this could exceed 255 by the largest bots.

+++ twilight-gateway-queue/src/in_memory.rs
@@ -27,13 +27,13 @@ enum Message {
 struct Settings {
     /// The maximum number of concurrent permits to grant. `0` instantly grants
     /// all permits.
-    max_concurrency: u8,
+    max_concurrency: u16,
     /// Remaining daily permits.
-    remaining: u16,
+    remaining: u32,
     /// Time until the daily permits reset.
     reset_after: Duration,
     /// The number of permits to reset to.
-    total: u16,
+    total: u32,
 }
 
 /// [`InMemoryQueue`]'s background task runner.
@@ -164,7 +164,7 @@ impl InMemoryQueue {
     /// # Panics
     ///
     /// Panics if `total` < `remaining`.
-    pub fn new(max_concurrency: u8, remaining: u16, reset_after: Duration, total: u16) -> Self {
+    pub fn new(max_concurrency: u16, remaining: u32, reset_after: Duration, total: u32) -> Self {
         assert!(total >= remaining);
         let (tx, rx) = mpsc::unbounded_channel();
 
@@ -222,7 +222,7 @@ impl InMemoryQueue {
     /// Panics if `total` < `remaining`.
     ///
     /// [Get Gateway Bot]: https://discord.com/developers/docs/topics/gateway#get-gateway-bot
-    pub fn update(&self, max_concurrency: u8, remaining: u16, reset_after: Duration, total: u16) {
+    pub fn update(&self, max_concurrency: u16, remaining: u32, reset_after: Duration, total: u32) {
         assert!(total >= remaining);
 
         self.tx
